requirements.txt
oandapyV20
pandas
python-dotenv
numpy
